Added series-selection annotation
Image IDs can now be annotated with '@' symbols to select a specific
series index. The syntax is:

* <ID>@<series#>@<.ext>

For example: test@27@.jpeg would open the 27th series of test.jpeg

The @<series#>@ can appear anywhere in the id, but if it appears in the
extension the image will probably not be supported correctly.

So te@27@st.jpeg is ok and would also open the 27th series of test.jpeg,
but test.jp@27@eg is not supported.