Corrected writer handshaking
The Write method was handshaking one too many times, leaving bytes in
the buffer.

Note that endCommand was updated to print a single endline again, to
avoid putting an excess of bytes on the buffer (in cases where the pipe
length may land between a 2nd and 3rd newline).

Also updated the handshaking method to clarifiy if it is reading more or
finished reading the bytes for a given plane.